requirements.txt
aiohttp==3.7.3
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
cryptography==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-import-export==2.5.0
django-password-reset==2.0
django-simpleui==2021.1.1
et-xmlfile==1.0.1
fake-useragent==0.1.11
geoip2==4.1.0
idna==2.10
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
maxminddb==2.0.3
mccabe==0.6.1
multidict==5.1.0
mysql-connector-python==8.0.22
odfpy==1.4.1
openpyxl==3.0.5
parse==1.18.0
protobuf==3.14.0
pycodestyle==2.6.0
pycparser==2.20
pylint==2.5.3
PyMySQL==0.10.1
PySocks==1.6.8
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.6.3
